0.2.6
- Fixed Progress bar to function by precentages instead of integers.

- Files are stored in memory as unsigned long long's, instead of regular integers to support large transfers.

- Large transfers of file are split by 300 Mb chunks.

- Fixed the server's directory string that is being returned by serializing it. ( no longer possible to confuse the clients Json parser with special characters. )

- Output window is now a text browser (clickable hyperlinks and selectable)

+++ OpenFTP Client/headers/file.h
@@ -5,7 +5,7 @@ class File
 {
 
 public:
-	File(QString _fileName, QString _filePath, int _fileSize, bool _isDir, QString _lastModified ,QPixmap _icon = {})
+	File(QString _fileName, QString _filePath, quint64 _fileSize, bool _isDir, QString _lastModified ,QPixmap _icon = {})
 		: fileName{ _fileName }, filePath{ _filePath }, fileSize{ _fileSize }, isDir{ _isDir }, lastModified{ _lastModified }, icon{ _icon } {};
 
 	File() {};
@@ -19,7 +19,7 @@ public:
 	QString fileName;
 	QString filePath;
 	QString lastModified;
-	int fileSize = 0;
+	quint64 fileSize = 0;
 	bool isDir = false;
 	QPixmap icon;
 };
